+++ src/Modules/Core/DBConstants/Achievement/AchievementIDs.php
@@ -4,5 +4,6 @@ namespace Foodsharing\Modules\Core\DBConstants\Achievement;
 
 class AchievementIDs
 {
+    final public const int KAM_CERTIFICATE = 1;
     final public const int HYGIENE_CERTIFICATE = 4;
 }