+++ src/Permissions/StoreChainPermissions.php
@@ -5,8 +5,6 @@ namespace Foodsharing\Permissions;
 use Foodsharing\Lib\Session;
 use Foodsharing\Modules\Core\DBConstants\Foodsaver\Role;
 use Foodsharing\Modules\Core\DBConstants\Region\RegionIDs;
-use Foodsharing\Modules\Region\RegionGateway;
-use Foodsharing\Modules\Store\StoreGateway;
 use Foodsharing\Modules\StoreChain\StoreChainGateway;
 use Foodsharing\Modules\Unit\CurrentUserUnitsInterface;
 
@@ -15,8 +13,6 @@ class StoreChainPermissions
     public function __construct(
         private readonly Session $session,
         private readonly StoreChainGateway $gateway,
-        private readonly StoreGateway $storeGateway,
-        private readonly RegionGateway $regionGateway,
         private readonly CurrentUserUnitsInterface $currentUserUnits,
     ) {
     }
@@ -33,10 +29,7 @@ class StoreChainPermissions
 
     public function maySeeChainList(): bool
     {
-        return $this->session->mayRole(Role::STORE_MANAGER) ||
-            ($this->session->mayRole(Role::FOODSAVER) &&
-                $this->regionGateway->hasMember($this->session->id(), RegionIDs::STORE_CHAIN_GROUP)) ||
-            $this->storeGateway->isStoreTeamMemberOfStoreChainStore($this->session->id());
+        return $this->session->mayRole(Role::FOODSAVER);
     }
 
     public function mayCreateChain(): bool
@@ -56,25 +49,6 @@ class StoreChainPermissions
 
     public function maySeeChainStores($chainId): bool
     {
-        if ($this->session->mayRole(Role::FOODSAVER) &&
-            $this->regionGateway->hasMember($this->session->id(), RegionIDs::STORE_CHAIN_GROUP)) {
-            return true;
-        }
-
-        return $this->mayAdministrateStoreChain($chainId);
-    }
-
-    public function maySeeChainDetails($chainId = null): bool
-    {
-        if ($this->session->mayRole(Role::FOODSAVER) &&
-            $this->regionGateway->hasMember($this->session->id(), RegionIDs::STORE_CHAIN_GROUP)) {
-            return true;
-        }
-
-        if (empty($chainId)) {
-            return $this->mayAdministrateStoreChains();
-        }
-
         return $this->mayAdministrateStoreChain($chainId);
     }
 }